rename functions to evaluator
